Move test code to the same package as the main code
This allows package-private types and members to be tested.
