+++ test/com/tvl/util/ImmutableArrayBuilderTest.java
@@ -1,8 +1,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-package com.tvl.util.test;
+package com.tvl.util;
 
 import com.google.common.collect.Iterables;
-import com.tvl.util.ImmutableArrayList;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
