+++ test/com/tvl/util/ImmutablesTestBase.java
@@ -1,5 +1,5 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-package com.tvl.util.test;
+package com.tvl.util;
 
 import com.google.common.collect.Iterables;
 import java.text.Collator;
